+++ pyscriptjs/tests/py-unit/test_python_plugins.py
@@ -35,7 +35,7 @@ class TestPyMarkdown:
         console_mock = Mock()
         monkeypatch.setattr(py_markdown, "console", console_mock)
         config = "just a config"
-        interpreter = "just a runtime"
+        interpreter = "just an interpreter"
 
         py_markdown.plugin.configure(config)
         console_mock.log.assert_called_with("configuration received: just a config")